[ref-01] enhance readability in components and create new functions and hooks to separate responsibilities
