+++ src/components/typed-captions-panel/component.tsx
@@ -1,6 +1,7 @@
 import { BbbPluginSdk, DataChannelTypes, PluginApi } from 'bigbluebutton-html-plugin-sdk';
 import { defineMessages, IntlShape } from 'react-intl';
 import * as React from 'react';
+import * as ReactDOM from 'react-dom/client';
 
 import Styled from './styles';
 import { TypedCaptionsInput } from './input-captions/component';
@@ -46,4 +47,22 @@ export function TypedCaptionsSidekickArea(props: GenericContentExampleProps) {
       />
     </Styled.CaptionsWrapper>
   );
+}
+
+export const renderComponent = (
+  uuid: string,
+  intl: IntlShape,
+  captionLocale: string,
+) => (element: HTMLElement) => {
+  const root = ReactDOM.createRoot(element);
+  root.render(
+    <React.StrictMode>
+      <TypedCaptionsSidekickArea
+        captionLocale={captionLocale}
+        uuid={uuid}
+        intl={intl}
+      />
+    </React.StrictMode>,
+  );
+  return root;
 }
