+++ src/components/typed-captions-panel/input-captions/component.tsx
