+++ src/components/typed-captions-panel/input-captions/styles.ts
