+++ src/components/typed-captions-panel/styles.ts
