+++ src/main/types.ts
@@ -1,3 +1,5 @@
+import { ActiveCaptionMenuInformation } from '../common/types';
+
 interface TypedCaptionsProps {
     pluginName: string,
     pluginUuid: string,
@@ -11,5 +13,7 @@ interface ExternalVideoMeetingSubscription {
         }
     }[]
 }
+
+export type PushActiveCaptionFunction = (args: ActiveCaptionMenuInformation) => void;
 
 export { TypedCaptionsProps, ExternalVideoMeetingSubscription };
